move methods from testnode to renter.go
